Refactor: extract EngineSettingsViewModel from SettingsViewModel (extract-and-delegate)
SettingsViewModel had grown to 2,568 lines and its documented decomposition
(SettingsRootViewModel.swift:24-28, SettingsView.swift) had stalled for lack of a
test net. With the engine surface now pinned by characterization tests, extract
the best-bounded slice.

- Add Sources/MacParakeetViewModels/EngineSettingsViewModel.swift
  (@MainActor @Observable, imports only Foundation + MacParakeetCore): owns engine
  preference + model variants, the switch-confirmation state machine, model
  status/availability/error state, downloads, repair, deletion, and the injected
  engine seams — moved verbatim, including the load-bearing isApplying* reentrancy
  guards and didSet persistence.
- SettingsViewModel now holds `public let engine` and forwards the full public
  engine surface (42 shims: 27 property/computed + 15 function/static, including
  the static speechEngineSwitchUnavailableMessage(for:)). Existing views and tests
  keep compiling unchanged — Sources/MacParakeet/Views is untouched.
- SettingsViewModel: 2,568 -> 1,733 lines (-835). The engine concern is now an
  isolated, independently testable unit.
- Add Tests/.../EngineSettingsViewModelTests.swift (17 tests) modeled on
  LLMSettingsViewModelTests.

Behavior-preserving: the committed regression suites pass with ZERO assertion
edits — SettingsViewModelTests (118) and SettingsEngineCharacterizationTests (19)
— which is the definition of a correct extract-and-delegate. Realizes the
EngineSettingsViewModel row in 2026-04-settings-ia-overhaul.md §3; re-pointing
views to bind engine.* directly is a deferred follow-up.

Executes plans/active/2026-06-15-settings-engine-viewmodel-extraction.md (plan
4/5, the gated extraction). Implemented by Codex (GPT-5.5 xhigh); independently
verified: swift build, both regression suites unchanged, EngineSettingsViewModel
tests (17), full swift test (3854, 0 failures), and the Swift 6 language-mode build.
