+++ plans/README.md
@@ -53,7 +53,7 @@
 | [2026-06-15-dx-format-lint-baseline](active/2026-06-15-dx-format-lint-baseline.md) | swift-format + .editorconfig + dev scripts + informational CI | **VERIFY-THEN-ARCHIVE** | P2 | Implemented in this branch: `.editorconfig`, `.swift-format`, opt-in format/check scripts, non-blocking CI lint, and AGENTS inner-loop note. Confirm merge, then archive. |
 | [2026-06-15-settings-observer-fanout-collapse](active/2026-06-15-settings-observer-fanout-collapse.md) | Table-driven AppSettingsObserverCoordinator | **VERIFY-THEN-ARCHIVE** ✅ | P2 | Implemented in this branch: 10 plain setting-change observers now register through a table-driven loop + token array, with onboarding/settings payload observers still explicit. Verified with `swift build`, focused coordinator tests, full `swift test`, Swift 6 no-Whisper build, and structural counts. Confirm commit/merge, then archive. |
 | [2026-06-15-settings-engine-characterization-tests](active/2026-06-15-settings-engine-characterization-tests.md) | Characterize the Settings engine/model surface | **VERIFY-THEN-ARCHIVE** ✅ | P2 | Implemented in this branch: 19 characterization tests pin selection defaults/persistence, switch-confirmation state, unavailable-message copy, and injected model-cache refresh behavior for the `SettingsViewModel` engine slice. Verified with `swift build`, focused new/existing Settings filters, and full `swift test`. |
-| [2026-06-15-settings-engine-viewmodel-extraction](active/2026-06-15-settings-engine-viewmodel-extraction.md) | Extract EngineSettingsViewModel (extract-and-delegate) | **TODO** (gated) | P2 | Net-new. L/MED. Executes the stalled `EngineSettingsViewModel` slice from `2026-04-settings-ia-overhaul.md` §3, test-first via extract-and-delegate. **HARD depends on** the characterization tests. Supersedes the Engine portion of the IA plan's god-file remainder. |
+| [2026-06-15-settings-engine-viewmodel-extraction](active/2026-06-15-settings-engine-viewmodel-extraction.md) | Extract EngineSettingsViewModel (extract-and-delegate) | **VERIFY-THEN-ARCHIVE** ✅ | P2 | Implemented in this branch: `EngineSettingsViewModel` owns the engine/model state, status refresh, downloads, repair, deletion, and switch-confirmation logic; `SettingsViewModel` holds `engine` and forwards the existing public surface, including the static unavailable-message helper. Existing views remain untouched. |
 | [2026-06-15-transcript-formatter-dedup](active/2026-06-15-transcript-formatter-dedup.md) | Shared TranscriptFormatter (kill the dup AI-formatter path) | **VERIFY-THEN-ARCHIVE** ✅ | P2 | Implemented in this branch: shared `TranscriptFormatter` + unified `FormatterOutcome`, Dictation/Transcription delegate to it (services −145 lines), 8 focused formatter tests added. Post-review tightening (Codex + Gemini converged): the five per-caller differences collapse into a typed `Lane` enum (4-param `format()`) so impossible states are unrepresentable. Capture-telemetry-emitter dedup remains deferred as a follow-up. |
 
 > [`active/2026-06-12-advisor-index.md`](active/2026-06-12-advisor-index.md) and
